fix: frontend integration update
