+++ frontend/src/main.tsx
@@ -7,11 +7,6 @@ import { AuthProvider } from "./contexts/AuthContext";
 
 import "./index.css";
 
-// if (import.meta.env.NODE_ENV === "development") {
-// 	const worker = require("./mocks/browser");
-// 	worker.start({ onUnhandledRequest: "bypass" });
-// }
-
 const router = createBrowserRouter(ROUTES);
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
