--- frontend/src/mocks/browser.ts
@@ -1,5 +1,0 @@
-// import { setupWorker } from "msw";
-// import { handlers } from "./handlers";
-
-// export const worker = setupWorker(...handlers);
-export const nil = null;
